requirements-dev.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.2; python_version >= '3.10'
djangorestframework==3.16.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.13.2; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
whitenoise==6.9.0; python_version >= '3.9'
